extracted inner class & enum into opibuilder.util; code cleanup;
