+++ applications/opibuilder/opibuilder-plugins/org.csstudio.opibuilder/src/org/csstudio/opibuilder/editparts/IPVWidgetEditpart.java
@@ -60,7 +60,7 @@ public interface IPVWidgetEditpart {
 
     public boolean isPVControlWidget();
 
-    public void performBeastBlink();
+    public void performBeastBlink(final int blinkState);
 
     public void resetBeastBlink();
 }
